BF 'long' is not a portable data type. Use int32_t

+++ iModelCore/DgnPlatform/DgnGeoCoord/DgnGeoCoord.cpp
@@ -2638,8 +2638,8 @@ const CSDefinition  &csDef
     CSMap::CS_stncp (params.ell_knm, csDef.elp_knm, DIM(params.ell_knm));
     DomainFromCsDef (params.gcDom, csDef);
 
-    params.zoneNo       =  (long) csDef.prj_prm1;
-    params.hemisphere   =  (long) (0 > csDef.prj_prm2 ?  OBE_HEMISPHERE_SOUTH : OBE_HEMISPHERE_NORTH);
+    params.zoneNo       =  (int32_t) csDef.prj_prm1;
+    params.hemisphere   =  (int32_t) (0 > csDef.prj_prm2 ?  OBE_HEMISPHERE_SOUTH : OBE_HEMISPHERE_NORTH);
 
 
     params.paper_scl    =  csDef.map_scl;
@@ -3423,7 +3423,7 @@ const CSDefinition  &csDef
     CSMap::CS_stncp (params.ell_knm, csDef.elp_knm, DIM(params.ell_knm));
     DomainFromCsDef (params.gcDom, csDef);
 
-    params.zoneNo       = (long) csDef.prj_prm1;
+    params.zoneNo       = (int32_t) csDef.prj_prm1;
     params.paper_scl    = csDef.map_scl;
     params.x_off        = csDef.x_off;
     params.y_off        = csDef.y_off;
